yash-unified-mdoc-user-story-a2a-server-dev/pyproject.toml
[project]
name = "mdoc-user-story"
version = "0.1.0"
description = "MDOC USER STORY"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "langchain-mcp-adapters>=0.1.8",
    "fastmcp>=2.10.6",
    "confluent-kafka",
    "fastapi",
    "kafka-python",
    "langchain-litellm>=0.3.0",
    'mangum',
    "boto3>=1.40.74",
    "psycopg2-binary>=2.9.11",
]

[tool.hatch.build.targets.wheel]
packages = ["common", "hosts", "agents"]

[tool.uv.workspace]
members = ["agents/crewai"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.2",
]
